python_service/requirements.txt
flask
psycopg2-binary
flask-cors
pydantic
flake8
black
